# backend/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API para o projeto MVP Scout Técnico",
    "version": "1.0.0",
    "title": "MVP Scout Técnico API"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/users": {
      "get": {
        "summary": "Retorna todos os usuários",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "summary": "Atualiza um usuário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso"
          }
        }
      },
      "delete": {
        "summary": "Deleta um usuário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso"
          }
        }
      }
    },
    "/players": {
      "get": {
        "summary": "Retorna todos os jogadores",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de jogadores",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Player"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adiciona um novo jogador",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Jogador adicionado com sucesso"
          }
        }
      }
    },
    "/players/{id}": {
      "put": {
        "summary": "Atualiza um jogador existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jogador atualizado com sucesso"
          }
        }
      },
      "delete": {
        "summary": "Deleta um jogador existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Jogador deletado com sucesso"
          }
        }
      }
    },
    "/teams": {
      "get": {
        "summary": "Retorna todos os times",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de times",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adiciona um novo time",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Team"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Time adicionado com sucesso"
          }
        }
      }
    },
    "/teams/{id}": {
      "put": {
        "summary": "Atualiza um time existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Team"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time atualizado com sucesso"
          }
        }
      },
      "delete": {
        "summary": "Deleta um time existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Time deletado com sucesso"
          }
        }
      }
    },
    "/coaches": {
      "get": {
        "summary": "Retorna todos os treinadores",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de treinadores",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Coach"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adiciona um novo treinador",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Coach"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Treinador adicionado com sucesso"
          }
        }
      }
    },
    "/coaches/{id}": {
      "put": {
        "summary": "Atualiza um treinador existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Coach"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Treinador atualizado com sucesso"
          }
        }
      },
      "delete": {
        "summary": "Deleta um treinador existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Treinador deletado com sucesso"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Player": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "position": {
          "type": "string"
        },
        "team": {
          "type": "string"
        },
        "goals": {
          "type": "integer"
        },
        "assists": {
          "type": "integer"
        }
      }
    },
    "Team": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "coach": {
          "type": "string"
        }
      }
    },
    "Coach": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "team": {
          "type": "string"
        }
      }
    }
  }
}